Move the SOE verification to a JVM specific test specification
